Early exit if playlist does not begin with #EXTM3U
Every HLS playlist MUST start with #EXTM3U. In the future we will add
support for other types of input (e.g. SCTE35 message, WebVTT URL, fMP4
URL, etc.) and when we do we will need to take the response content-type
into account to switch on what parsing we make.