Got segment view for WebVTT working!
Again, lots of rough edges, but happy at the progress. I think the next
thing I'd like to do is mp4 view. That may be tricky but I believe I
will learn a lot about the direction that everything else needs to go in
based on how I solve that.