Minor styling change
I think this better distributes the views within the viewer when the
supplemental is open. I noticed that in the Apple example HLS stream
that the copy button covered the target duration on the subtitles
playlist when the screen was shrank small enough. Mainly because the
playlist is so narrow (short URLs), so this helps keep that from
happening.