Path included in routing
Annoyingly, the Leptos router doesn't seem to work well with the fact
that GitHub Pages publishes the site to a path on the main domain, so
all the paths defined in the router get messed up... Even with the trunk
public url declaration.

Ideally I'd like this to be more reactive to the environment... I can
get the base_uri from the Document; hoewever, the paths in the router
need to be defined with &'static str, so my hands are tied a little...
Perhaps I could try something with leaking a string... But for now I
just want to get the site to work.