fix!: updated separator used for definition lists
Changed from using `%0A` to `%22` when encoding lists of definitions
into the URL query string, based on this issue:
https://github.com/leptos-rs/leptos/issues/4232

This is just a quick fix for now. I will revisit this with a value
encoding strategy such that we can choose a safe separator character
that can be guaranteed to be encoded away from any value coming from the
playlist. This will add an extra decoding step so is a bit more of a
change then just switching out the character used for the separator.

BREAKING CHANGE: This changes the URL format, so anyone that may have
saved a URL from before this change, and if that URL also included a
list of parsed definitions from the playlist, then they will find that
the URL no longer works after this change. For now things are still
pretty fluid and there are no users yet, so I'm happy to just break. But
in the future I may consider introducing a URL versioning scheme to
better support backwards compatibility.