feat: updated link tag construction in playlist
The `PlaylistViewer` now creates link tags (tags that have attributes
that link to playlists or segments) using `UnknownTag` and iterating
through the attributes. This seems more robust as it does not rely on
fragile string find, split, and replace methods on the inner tag, at the
expense of having the attributes already parsed for us. We make use of
the new `TagValue::try_as_ordered_attribute_list` method so that we can
iterate through the attributes and put the anchor tag around the value
we care about.

This is helpful in preparation for SCTE35 support, where the daterange
tag can have multiple splits (SCTE35-OUT, SCTE35-IN, SCTE35-CMD), and so
extending the existing logic would've been too difficult.