feat: support for scte35 href
This adds the mechanics for encoding and decoding the SCTE35 state with
the URL. Methods for getting the href for a SCTE35 message were added
and also the functionality to take a string and decode it into a
`Scte35Context` were added too. Next step is to utilize those with links
from the playlist and a dedicated SCTE35 supplemental context view.