feat: views for new atoms and defend against crash
As well as adding views for the new atoms, I added a Clippy deny lint
for `wildcard_enum_match_arm`, which forces all known arms of the `Any`
enum to be matched against (which is helpful because it is declared as
`#[non_exhaustive]` so we can't check all cases). The deploy job was
also updated to check for Clippy failures before attempting to publish
anything so another extra bit of protection from pushing something that
crashes.