fix: relative path map uri highlights correctly
I was previously checking if the URL derived from the query parameter
contained the EXT-X-MAP:URI. This does not work if the relative path has
"dot dot" path components _(going upwards in directories)_.

I considered making an equality check on discontinuity sequence, but I
don't think it is required to introduce a new map for a new
discontinuity (though it is common).

I considered leaving it with just the media sequence check, but I don't
like that because returning to an old "viewer URL" on a live playlist
that has moved on a lot and has a new map tag will falsely highlight the
new one in that case.

So the solution I went with is to resolve to absolute the URI that I'm
checking before comparing against the URL taken from the query parameter
which is going to be absolute as well. This should then be an equality
check.

Fixes #14