refactor: moved atom properties to separate files
The mp4_atom_properties file was way too long. There was a massive match
statement where each atom was converted into AtomProperties that are
used to carry the data for display in the UI. The change moves each impl
into its own dedicated file for that atom, which I believe is cleaner
and easier to reason about / read.