feat: senc box parsing
Added support for parsing and displaying senc boxes. This was a little
tricky, because the information I need to decode the box properly does
not actually exist in the box (not even the file, given it is in the
tenc within the init section). The tactic I went with is to restrict the
supported sizes of IV to 0, 8, or 16, and then try each, until I can
completely decode the correct number of bytes.